Moves StoredAccountMeta tests to append_vec::meta module
