+++ accounts-db/src/append_vec/meta.rs
@@ -211,3 +211,48 @@ impl IsZeroLamport for StoredAccountNoData<'_> {
         self.lamports() == 0
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use {
+        super::*,
+        solana_account::{accounts_equal, Account},
+    };
+
+    #[test]
+    fn test_stored_readable_account() {
+        let lamports = 1;
+        let owner = Pubkey::new_unique();
+        let executable = true;
+        let rent_epoch = 2;
+        let meta = StoredMeta {
+            write_version_obsolete: 5,
+            pubkey: Pubkey::new_unique(),
+            data_len: 7,
+        };
+        let account_meta = AccountMeta {
+            lamports,
+            owner,
+            executable,
+            rent_epoch,
+        };
+        let data = Vec::new();
+        let account = Account {
+            lamports,
+            owner,
+            executable,
+            rent_epoch,
+            data: data.clone(),
+        };
+        let offset = 99 * size_of::<u64>(); // offset needs to be 8 byte aligned
+        let stored_size = 101;
+        let stored_account = StoredAccountMeta {
+            meta: &meta,
+            account_meta: &account_meta,
+            data: &data,
+            offset,
+            stored_size,
+        };
+        assert!(accounts_equal(&account, &stored_account));
+    }
+}
